refactor: Update ImageKitLoader import path and add error, not-found components
